Misc Bug Fixes + Some more refactoring...
